refactor : structure list provider
