+++ lib/page/home/home_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
 
 import 'package:lister/page/search/search_page.dart';
-import '../../model/provider/main_lists_provider.dart';
+import '../../model/provider/lists_provider.dart';
 import '../../widget/custom/custom_search_bar.dart';
 import '../../widget/list/home_list_view.dart';
 
@@ -122,8 +122,8 @@ class HomePage extends StatelessWidget {
 
                 SizedBox(
                   height: 282,
-                  child: ChangeNotifierProvider<MainListsProvider>(
-                    create: (context) => MainListsProvider(),
+                  child: ChangeNotifierProvider<GetListsProvider>(
+                    create: (context) => GetListsProvider(),
                     child: HomeListView(),
                   ),
                 )
