+++ lib/widget/custom/custom_book_mark_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
 
-import '../../model/provider/main_lists_provider.dart';
+import '../../model/provider/lists_provider.dart';
 
 class CustomBookMarkButton extends StatelessWidget {
   final int index;
@@ -19,7 +19,7 @@ class CustomBookMarkButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<MainListsProvider>(
+    return Consumer<GetListsProvider>(
       builder: (context, provider, child) {
         return Positioned(
           top: 4.0.h,
